+++ fpy2/__init__.py
@@ -45,19 +45,15 @@ from .backend import (
 )
 
 from .interpret import (
+    Interpreter,
     PythonInterpreter,
     RealInterpreter,
     TitanicInterpreter,
-)
-
-from .runtime import (
-    Function,
-    Interpreter,
     set_default_interpreter,
     get_default_interpreter,
-    FunctionProfiler,
-    ExprProfiler,
 )
+
+from .runtime import Function
 
 from .utils import (
     fraction,
