+++ fpy2/interpret/__init__.py
@@ -1,5 +1,9 @@
 """Interpreters for FPy."""
 
+from .interpreter import Interpreter, get_default_interpreter, set_default_interpreter
+
 from .native import PythonInterpreter
 from .real import RealInterpreter
 from .titanic import TitanicInterpreter
+
+set_default_interpreter(TitanicInterpreter())
