+++ fpy2/profile/function_profiler.py
@@ -9,10 +9,11 @@ from typing import Any, Optional
 from titanfp.arithmetic.ieee754 import Float, IEEECtx
 from titanfp.arithmetic.mpmf import MPMF
 
-from ..function import Function, Interpreter, get_default_interpreter
-from ..metric import ordinal_error
-from ...interpret.real import RealInterpreter
-from ..real.rival_manager import PrecisionLimitExceeded
+from ..interpret import Interpreter, RealInterpreter, get_default_interpreter
+from ..runtime import Function
+from ..runtime.metric import ordinal_error
+from ..runtime.real import PrecisionLimitExceeded
+
 
 class FunctionProfileResult:
     """
