+++ fpy2/runtime/__init__.py
@@ -1,13 +1,6 @@
 from .env import ForeignEnv
 from .function import (
     Function,
-    Interpreter,
-    get_default_interpreter,
-    set_default_interpreter
+    get_default_function_call,
+    set_default_function_call,
 )
-
-from ..interpret import TitanicInterpreter
-
-from .profile import ExprProfiler, FunctionProfiler
-
-set_default_interpreter(TitanicInterpreter())
