+++ fpy2/runtime/real/__init__.py
@@ -1,1 +1,2 @@
 from .interval import Interval, BoolInterval, RealInterval
+from .rival_manager import RivalManager, PrecisionLimitExceeded
